refactor into command line legatest
