+++ requirements.txt
@@ -5,6 +5,6 @@ paramiko
 minio
 PyYAML
 requests
-git+https://github.com/NBISweden/LocalEGA-cryptor.git
+git+https://github.com/neicnordic/LocalEGA-cryptor.git
 psycopg2-binary
 tenacity
